Checked template applying function for different inputs (see issue #32) added error handling for template applying added file path checks for file deletions

+++ src/boolean_model/views.py
@@ -13,26 +13,10 @@ from quick_format.models import Quick
 from quick_format.views import quick_detail
 import pickle
 from rxncon_system.models import Rxncon_system
-
-import rxncon.input.excel_book.excel_book as rxncon_excel
-import rxncon.input.quick.quick as rxncon_quick
-import rxncon.simulation.boolean.boolean_model as rxncon_boolean_model
-import rxncon.simulation.boolean.boolnet_from_boolean_model as bfbm
-from rxncon.simulation.boolean.boolean_model import boolean_model_from_rxncon, \
-    SmoothingStrategy, KnockoutStrategy, OverexpressionStrategy
+from rxncon.simulation.boolean.boolean_model import SmoothingStrategy, KnockoutStrategy, OverexpressionStrategy
 from rxncon.simulation.boolean.boolnet_from_boolean_model import QuantitativeContingencyStrategy, boolnet_strs_from_rxncon
 
 
-
-# def create_rxncon_system(system, system_type):
-#     if system_type == "File":
-#         try:
-#             book = rxncon_excel.ExcelBook(system.get_absolute_path())
-#         except:
-#             book = rxncon_excel.ExcelBook(system.get_absolute_path())
-#     else:
-#         book = rxncon_quick.Quick(system.quick_input)
-#     return book.rxncon_system
 
 
 def check_filepath(request, file_path, file, media_root):
@@ -147,9 +131,10 @@ def bool_delete(request, pk):
         form = DeleteBoolForm(request.POST, instance=f)
 
         if form.is_valid(): # checks CSRF
-            os.remove(f.model_path.name)
-            os.remove(f.symbol_path.name)
-            os.remove(f.init_path.name)
+            if os.path.exists(f.model_path.name):
+                os.remove(f.model_path.name)
+                os.remove(f.symbol_path.name)
+                os.remove(f.init_path.name)
             f.delete()
             messages.success(request, "Tree files Successfully deleted.")
             if system_type == "Quick":
